+++ src/services/webhooks/WebhooksService.test.ts
@@ -1,5 +1,5 @@
 import * as sinon from 'sinon';
-import * as Qminder from '../../src/qminder-api';
+import * as Qminder from '../../qminder-api';
 
 describe('Qminder.webhooks', function () {
   let requestStub: sinon.SinonStub;
