--- test/unit/jestSetupFile.ts
@@ -1,1 +1,0 @@
-(global as any).VERSION = 'unit-tests';
